+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Next Version
 
 * `@ContributesBinding` supports qualifiers now, see the README and documentation for examples.  
-* Upgrade Dagger to `2.32`. Generating factories for assisted injection is no longer compatible with older Dagger versions due to the behavior change in Dagger itself. Make sure to use Dagger version `2.32` or newer in your project, too.  
+* Upgrade Dagger to `2.32`. Generating factories for assisted injection is no longer compatible with older Dagger versions due to the behavior change in Dagger itself. Make sure to use Dagger version `2.32` or newer in your project, too.
+* Use the mangled function name to generate the factory for a provider method.
 
 ## 2.1.0 (2021-02-05)